requirements.txt
python-dotenv>=1.0.0
requests>=2.28.0
finnhub-python>=2.4.25
yfinance>=0.2.25
pycoingecko>=2.1.0
fastapi>=0.95.0
uvicorn>=0.22.0
pytest>=7.2.0
uagents>=0.8.1
